Change makefile naming to be consistent for gpMgmt/bin
Previously, the "all" target was used only by gpMgmt/bin/Makefile in
place of the "install" target used elsewhere. The underlying reason was
because python doesn't need a compile phase (the "all" target).

However, this is inconsistent. Change to use the standard "install"
target, which is also the default on the gpMgmt/bin/Makefile.

Signed-off-by: Larry Hamel <lhamel@pivotal.io>